backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
pydantic
pydantic-settings
